move custom_domain into domain
